Update AsyncWebSocket.cpp
Fix compile error for ESP32-C3 toolchain ( ambiguous type)

+++ src/AsyncWebSocket.cpp
@@ -829,7 +829,7 @@ void AsyncWebSocketClient::binary(AsyncWebSocketMessageBuffer * buffer)
 
 IPAddress AsyncWebSocketClient::remoteIP() {
     if(!_client) {
-        return IPAddress(0U);
+        return IPAddress((uint32_t)0);
     }
     return _client->remoteIP();
 }
